+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.43.34](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.33...@synapsecns/synapse-interface@0.43.34) (2025-11-28)
+
+
+### Bug Fixes
+
+* **synapse-interface:** fix navigation menu error ([#3863](https://github.com/synapsecns/sanguine/issues/3863)) ([15cd050](https://github.com/synapsecns/sanguine/commit/15cd050dbc9a6cbfb9dcc70bb2686948e258d851))
+
+
+
+
+
+## [0.43.33](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.32...@synapsecns/synapse-interface@0.43.33) (2025-11-27)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.43.32](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.31...@synapsecns/synapse-interface@0.43.32) (2025-11-27)
+
+
+### Bug Fixes
+
+* **synapse-interface:** refetch quote on address changes ([#3849](https://github.com/synapsecns/sanguine/issues/3849)) ([a7f47ec](https://github.com/synapsecns/sanguine/commit/a7f47ec8a664042399a7bf6676d1a38724e6e417))
+
+
+
+
+
+## [0.43.31](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.30...@synapsecns/synapse-interface@0.43.31) (2025-11-26)
+
+
+### Bug Fixes
+
+* replace RPCs that are not responding ([#3846](https://github.com/synapsecns/sanguine/issues/3846)) ([1c67df0](https://github.com/synapsecns/sanguine/commit/1c67df040dec29cf2a16cd427226c60ef7f1bd2b))
+
+
+
+
+
+## [0.43.30](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.29...@synapsecns/synapse-interface@0.43.30) (2025-11-25)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.43.29](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.28...@synapsecns/synapse-interface@0.43.29) (2025-11-24)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.43.28](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.43.27...@synapsecns/synapse-interface@0.43.28) (2025-11-13)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface